move some func
