+++ GameControl/GlobalAPI/openOrCloseContainer.go
@@ -36,33 +36,6 @@ func (g *GlobalAPI) OpenContainer(
 }
 
 /*
-打开背包。
-返回值的第一项代表执行结果，为真时背包被成功打开，否则反之。
-
-请确保打开前占用了容器资源，否则会造成程序惊慌。
-*/
-func (g *GlobalAPI) OpenInventory() (bool, error) {
-	g.Resources.Container.AwaitChangesBeforeSendPacket()
-	// await responce before send packet
-	err := g.WritePacket(&packet.Interact{
-		ActionType:            packet.InteractActionOpenInventory,
-		TargetEntityRuntimeID: g.BotInfo.BotRunTimeID,
-	})
-	if err != nil {
-		return false, fmt.Errorf("OpenInventory: %v", err)
-	}
-	// open inventory
-	g.Resources.Container.AwaitChangesAfterSendPacket()
-	// wait changes
-	if g.Resources.Container.GetContainerOpenDatas() == nil {
-		return false, nil
-	}
-	// if unsuccess
-	return true, nil
-	// return
-}
-
-/*
 关闭已经打开的容器，且只有当容器被关闭后才会返回值。
 您应该确保容器被关闭后，对应的容器公用资源被释放。
